Remove unnecessary parameter.
It had been my intention to add a "joy_stick_reopen()" call to the API,
which would allow the user to reopen a joystick device to a different
device node.

That would have caused too much problems, however, so I decided to
remove it from the public API. We still use it, but a parameter that is
always passed as NULL isn't very useful.

We still pass the other parameter as NULL. Should change that, but
that'd be by actually passing a GError there, rather than removing the
parameter.